refactored and eliminated old package
